+++ CHANGELOG.md
@@ -3,3 +3,4 @@ CHANGELOG for Sulu Sales
 
 * dev-develop
     * FEATURE     [CoreBundle]    Added handling of gross-prices to item-table
+    * FEATURE     [CoreBundle]    Added recurring price for items